orders/src/routes/new.ts
import mongoose from 'mongoose'
import express, {Request, Response} from 'express';
import { NotFoundError, OrderStatus, requireAuth, validateRequest, BadRequestError } from '@smax_ticketing/common';
import { body } from 'express-validator';
import { Ticket } from '../models/tickets';
import { Order } from '../models/order';
import { OrderCreatedPublisher } from '../events/publishers/order-created-publisher';
import { natsWrapper } from '../nats-wrapper';

const router = express.Router();

const EXPIRATION_WINDOW_SECONDS = 15 * 60;

router.post('/api/orders', requireAuth, [
    body('ticketId')
        .not()
        .isEmpty()
        .custom((input: string)=> mongoose.Types.ObjectId.isValid(input))
        .withMessage('TicketId must be provided')
], validateRequest, async (req: Request, res: Response)=>{
    // Find the tiсket the user is trying to order in the db
    const {ticketId} = req.body;
    const ticket = await Ticket.findById(ticketId);
    if (!ticket){
        throw new NotFoundError();
    }
    // Make sure this ticket is not arleady reserved
    // Run query to look at all orders. Find order where the ticket 
    // is the ticket we just found and the orders status is not cancelled
    // if we found, it means the ticket is reserved
    const isReserved = await ticket.isReserved();
    if(isReserved){
        throw new BadRequestError('Ticket is arleady reserved');
    }
    // Calculate an expiration date
    const expiration = new Date();
    expiration.setSeconds(expiration.getSeconds() + EXPIRATION_WINDOW_SECONDS)
    // Build the order and save it to the database
    const order = Order.build({
        userId: req.currentUser!.id,
        status: OrderStatus.Created,
        expiresAt: expiration,
        ticket: ticket
    })
    await order.save();
    // Publish an event saying an order was created
    new OrderCreatedPublisher(natsWrapper.client).publish({
        id: order.id,
        version: order.version,
        status: order.status,
        userId: order.userId,
        expiresAt: order.expiresAt.toISOString(),
        ticket: {
            id: ticket.id,
            price: ticket.price
        }
    })
    res.status(201).send(order);
})

export {router as newOrderRouter}
